Rework Drop Filter pick up logic
There was a bad interaction between the Voiding Drop Filter and a magnet
such as the Draconic Evolution Item Dislocator where it tries to pick up
a dropped item repeatedly in a loop. This changes the logic to mimic the
/dev/null from OpenBlocks, which instead of canceling the drop event,
just sets the item stack to 0. This seems to work better with magnets.