Voiding Drop Filter rework: tooltips, prevent looping pick up with magnet, fix crash moving filter while its GUI is opened (#2)
* spotless apply

* Update buildscript and gradle wrapper

* Rework Drop Filter pick up logic

There was a bad interaction between the Voiding Drop Filter and a magnet
such as the Draconic Evolution Item Dislocator where it tries to pick up
a dropped item repeatedly in a loop. This changes the logic to mimic the
/dev/null from OpenBlocks, which instead of canceling the drop event,
just sets the item stack to 0. This seems to work better with magnets.

* Describe Drop Filter functionality in tooltip

* Prevent Drop Filter from being moved while its GUI is opened

There are some additional checks that are necessary to prevent moving of
an item from the hotbar while its GUI is open. This is necessary for
some other items as well, so move the functionality into a new base
class.

* Update Github Actions workflow